Added Subjective calls to SoapeService

+++ backend/src/main/java/xyz/hidasy/rest/SoapeModel.java
@@ -7,15 +7,15 @@ import java.util.Date;
  */
 public class SoapeModel {
 
-    protected Integer appointmentId;
+    protected Long appointmentId;
 
     protected String lastUpdatedAt;
 
-    public Integer getAppointmentId() {
+    public Long getAppointmentId() {
         return appointmentId;
     }
 
-    public void setAppointmentId(Integer appointmentId) {
+    public void setAppointmentId(Long appointmentId) {
         this.appointmentId = appointmentId;
     }
